Create README.md
This project is a single-page web app where you can see 6 gifs. Upon clicking on those gifs, you will see the name of the mood and a brain fact about the mood. The DOM content will load first so as not to slow down the functionality. You will be able to like a gif to display your current mood.